✨ CGB boot rom

+++ ceres-core/src/boot_rom.rs
@@ -16,7 +16,7 @@ impl BootRom {
         }
     }
 
-    pub fn read(&self, address: u8) -> u8 {
+    pub fn read(&self, address: u16) -> u8 {
         self.boot_rom[address as usize]
     }
